Shuffle addresses when create cetcd client
cetcd use a round-robin method to failover the requests. However,
the round-robin is not global of all clients. All client on server A
would failover to server B if A is down because every client get the
next server maybe by the same address list. We address this by shuffling
the address list. Refer go-etcd: https://github.com/coreos/go-etcd/pull/231